Refactor module `keycloak` to `implementation`
Use existing Keycloak implementation as the base for the new generic, configuration driven implementation.
A custom publishing naeme for the artefact is also defined as `oidc` so that package is published as `app.quickcase.spring:oidc`
All references to Keycloak have been replaced by generic or default terms.
