+++ settings.gradle
@@ -1,3 +1,3 @@
 rootProject.name = 'spring-oidc'
 
-include 'api', 'keycloak'
+include 'api', 'implementation'
